+++ priv/static/phoenix_live_view.cjs.js
@@ -624,7 +624,9 @@ var DOM = {
     }
   },
   mergeFocusedInput(target, source) {
-    DOM.mergeAttrs(target, source, { exclude: ["value"] });
+    if (!(target instanceof HTMLSelectElement)) {
+      DOM.mergeAttrs(target, source, { exclude: ["value"] });
+    }
     if (source.readOnly) {
       target.setAttribute("readonly", true);
     } else {
@@ -635,6 +637,9 @@ var DOM = {
     return el.setSelectionRange && (el.type === "text" || el.type === "textarea");
   },
   restoreFocus(focused, selectionStart, selectionEnd) {
+    if (focused instanceof HTMLSelectElement) {
+      focused.focus();
+    }
     if (!DOM.isTextualInput(focused)) {
       return;
     }
@@ -1789,13 +1794,11 @@ var DOMPatch = class {
     liveSocket.time("morphdom", () => {
       this.streams.forEach(([ref, inserts, deleteIds, reset]) => {
         Object.entries(inserts).forEach(([key, [streamAt, limit]]) => {
-          this.streamInserts[key] = { ref, streamAt, limit, resetKept: false };
+          this.streamInserts[key] = { ref, streamAt, limit, reset };
         });
         if (reset !== void 0) {
           dom_default.all(container, `[${PHX_STREAM_REF}="${ref}"]`, (child) => {
-            if (inserts[child.id]) {
-              this.streamInserts[child.id].resetKept = true;
-            } else {
+            if (!inserts[child.id]) {
               this.removeStreamChildElement(child);
             }
           });
@@ -1875,12 +1878,10 @@ var DOMPatch = class {
         },
         onBeforeElChildrenUpdated: (fromEl, toEl) => {
           if (fromEl.getAttribute(phxUpdate) === PHX_STREAM) {
-            let toIds = Array.from(toEl.children).map((child) => child.id);
-            Array.from(fromEl.children).filter((child) => {
-              let { resetKept } = this.getStreamInsert(child);
-              return resetKept;
-            }).forEach((child) => {
-              this.streamInserts[child.id].streamAt = toIds.indexOf(child.id);
+            Array.from(toEl.children).forEach((child, idx) => {
+              if (this.streamInserts[child.id].reset) {
+                this.streamInserts[child.id].streamAt = idx;
+              }
             });
           }
         },
@@ -1911,6 +1912,7 @@ var DOMPatch = class {
           dom_default.maybeAddPrivateHooks(toEl, phxViewportTop, phxViewportBottom);
           dom_default.cleanChildNodes(toEl, phxUpdate);
           if (this.skipCIDSibling(toEl)) {
+            this.maybeReOrderStream(fromEl);
             return false;
           }
           if (dom_default.isPhxSticky(fromEl)) {
@@ -1946,7 +1948,8 @@ var DOMPatch = class {
           }
           dom_default.copyPrivates(toEl, fromEl);
           let isFocusedFormEl = focused && fromEl.isSameNode(focused) && dom_default.isFormInput(fromEl);
-          if (isFocusedFormEl && fromEl.type !== "hidden") {
+          let focusedSelectChanged = isFocusedFormEl && this.isChangedSelect(fromEl, toEl);
+          if (isFocusedFormEl && fromEl.type !== "hidden" && !focusedSelectChanged) {
             this.trackBefore("updated", fromEl, toEl);
             dom_default.mergeFocusedInput(fromEl, toEl);
             dom_default.syncAttrsToProps(fromEl);
@@ -1955,6 +1958,9 @@ var DOMPatch = class {
             trackedInputs.push(fromEl);
             return false;
           } else {
+            if (focusedSelectChanged) {
+              fromEl.blur();
+            }
             if (dom_default.isPhxUpdate(toEl, phxUpdate, ["append", "prepend"])) {
               appendPrependUpdates.push(new DOMPostMorphRestorer(fromEl, toEl, toEl.getAttribute(phxUpdate)));
             }
@@ -2051,6 +2057,20 @@ var DOMPatch = class {
         this.trackAfter("transitionsDiscarded", pendingRemoves);
       });
     }
+  }
+  isChangedSelect(fromEl, toEl) {
+    if (!(fromEl instanceof HTMLSelectElement) || fromEl.multiple) {
+      return false;
+    }
+    if (fromEl.options.length !== toEl.options.length) {
+      return true;
+    }
+    let fromSelected = fromEl.selectedOptions[0];
+    let toSelected = toEl.selectedOptions[0];
+    if (fromSelected && fromSelected.hasAttribute("selected")) {
+      toSelected.setAttribute("selected", fromSelected.getAttribute("selected"));
+    }
+    return !fromEl.isEqualNode(toEl);
   }
   isCIDPatch() {
     return this.cidPatch;
@@ -3426,6 +3446,7 @@ var View = class {
         if (disableText !== "") {
           el.innerText = disableText;
         }
+        el.setAttribute(PHX_DISABLED, el.disabled);
         el.setAttribute("disabled", "");
       }
     });