+++ priv/static/phoenix_live_view.js
@@ -653,7 +653,9 @@ var LiveView = (() => {
       }
     },
     mergeFocusedInput(target, source) {
-      DOM.mergeAttrs(target, source, { exclude: ["value"] });
+      if (!(target instanceof HTMLSelectElement)) {
+        DOM.mergeAttrs(target, source, { exclude: ["value"] });
+      }
       if (source.readOnly) {
         target.setAttribute("readonly", true);
       } else {
@@ -664,6 +666,9 @@ var LiveView = (() => {
       return el.setSelectionRange && (el.type === "text" || el.type === "textarea");
     },
     restoreFocus(focused, selectionStart, selectionEnd) {
+      if (focused instanceof HTMLSelectElement) {
+        focused.focus();
+      }
       if (!DOM.isTextualInput(focused)) {
         return;
       }
@@ -1818,13 +1823,11 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
       liveSocket.time("morphdom", () => {
         this.streams.forEach(([ref, inserts, deleteIds, reset]) => {
           Object.entries(inserts).forEach(([key, [streamAt, limit]]) => {
-            this.streamInserts[key] = { ref, streamAt, limit, resetKept: false };
+            this.streamInserts[key] = { ref, streamAt, limit, reset };
           });
           if (reset !== void 0) {
             dom_default.all(container, `[${PHX_STREAM_REF}="${ref}"]`, (child) => {
-              if (inserts[child.id]) {
-                this.streamInserts[child.id].resetKept = true;
-              } else {
+              if (!inserts[child.id]) {
                 this.removeStreamChildElement(child);
               }
             });
@@ -1904,12 +1907,10 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
           },
           onBeforeElChildrenUpdated: (fromEl, toEl) => {
             if (fromEl.getAttribute(phxUpdate) === PHX_STREAM) {
-              let toIds = Array.from(toEl.children).map((child) => child.id);
-              Array.from(fromEl.children).filter((child) => {
-                let { resetKept } = this.getStreamInsert(child);
-                return resetKept;
-              }).forEach((child) => {
-                this.streamInserts[child.id].streamAt = toIds.indexOf(child.id);
+              Array.from(toEl.children).forEach((child, idx) => {
+                if (this.streamInserts[child.id].reset) {
+                  this.streamInserts[child.id].streamAt = idx;
+                }
               });
             }
           },
@@ -1940,6 +1941,7 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
             dom_default.maybeAddPrivateHooks(toEl, phxViewportTop, phxViewportBottom);
             dom_default.cleanChildNodes(toEl, phxUpdate);
             if (this.skipCIDSibling(toEl)) {
+              this.maybeReOrderStream(fromEl);
               return false;
             }
             if (dom_default.isPhxSticky(fromEl)) {
@@ -1975,7 +1977,8 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
             }
             dom_default.copyPrivates(toEl, fromEl);
             let isFocusedFormEl = focused && fromEl.isSameNode(focused) && dom_default.isFormInput(fromEl);
-            if (isFocusedFormEl && fromEl.type !== "hidden") {
+            let focusedSelectChanged = isFocusedFormEl && this.isChangedSelect(fromEl, toEl);
+            if (isFocusedFormEl && fromEl.type !== "hidden" && !focusedSelectChanged) {
               this.trackBefore("updated", fromEl, toEl);
               dom_default.mergeFocusedInput(fromEl, toEl);
               dom_default.syncAttrsToProps(fromEl);
@@ -1984,6 +1987,9 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
               trackedInputs.push(fromEl);
               return false;
             } else {
+              if (focusedSelectChanged) {
+                fromEl.blur();
+              }
               if (dom_default.isPhxUpdate(toEl, phxUpdate, ["append", "prepend"])) {
                 appendPrependUpdates.push(new DOMPostMorphRestorer(fromEl, toEl, toEl.getAttribute(phxUpdate)));
               }
@@ -2080,6 +2086,20 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
           this.trackAfter("transitionsDiscarded", pendingRemoves);
         });
       }
+    }
+    isChangedSelect(fromEl, toEl) {
+      if (!(fromEl instanceof HTMLSelectElement) || fromEl.multiple) {
+        return false;
+      }
+      if (fromEl.options.length !== toEl.options.length) {
+        return true;
+      }
+      let fromSelected = fromEl.selectedOptions[0];
+      let toSelected = toEl.selectedOptions[0];
+      if (fromSelected && fromSelected.hasAttribute("selected")) {
+        toSelected.setAttribute("selected", fromSelected.getAttribute("selected"));
+      }
+      return !fromEl.isEqualNode(toEl);
     }
     isCIDPatch() {
       return this.cidPatch;
@@ -3455,6 +3475,7 @@ removing illegal node: "${(childNode.outerHTML || childNode.nodeValue).trim()}"
           if (disableText !== "") {
             el.innerText = disableText;
           }
+          el.setAttribute(PHX_DISABLED, el.disabled);
           el.setAttribute("disabled", "");
         }
       });